+++ lib/widgets/nk_button.dart
@@ -5,14 +5,21 @@ class NkButton extends StatelessWidget {
       {Key key,
       this.margin,
       this.color,
-      this.height,
+      this.height = 40,
       this.padding,
       this.onClick,
       this.textStyle,
-      this.width = 100,
+      this.width = 90,
       this.borderRadius = 10,
-      this.formKey})
-      : super(key: key);
+      this.formKey,
+      this.background})
+      : assert(
+            background == null || color == null,
+            'Impossivel usar uma cor com um widget de fundo\n'
+            'Escolha entre um widget de fundo ou uma cor.'),
+        assert(width >= 90, 'Tamanho mínimo excedido; minWidht = 90.'),
+        assert(height >= 40, 'Tamanho mínimo excedido; minHeight = 40.'),
+        super(key: key);
 
   final String text;
   final Color color;
@@ -24,36 +31,47 @@ class NkButton extends StatelessWidget {
   final GlobalKey<FormState> formKey;
   final EdgeInsetsGeometry margin;
   final EdgeInsetsGeometry padding;
+  final Background background;
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      height: height,
-      width: width,
-      child: Container(
-        margin: margin,
-        padding: padding,
-        child: RaisedButton(
-          onPressed: onClick,
-          shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(borderRadius)),
-          padding: EdgeInsets.all(0.0),
-          child: Ink(
-            decoration: BoxDecoration(
-                color: color,
-                borderRadius: BorderRadius.circular(borderRadius)),
-            child: Container(
-              constraints: BoxConstraints(maxWidth: width, minHeight: 50.0),
-              alignment: Alignment.center,
-              child: Text(
-                text,
-                textAlign: TextAlign.center,
-                style: textStyle,
-              ),
+    return Container(
+      constraints: BoxConstraints(
+          minHeight: 40.0, minWidth: 90, maxWidth: width, maxHeight: height),
+      margin: margin,
+      padding: padding,
+      child: RaisedButton(
+        onPressed: onClick,
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(borderRadius)),
+        padding: EdgeInsets.all(0.0),
+        child: Ink(
+          decoration: BoxDecoration(
+              color: color, borderRadius: BorderRadius.circular(borderRadius)),
+          child: Container(
+            width: width,
+            height: height,
+            constraints: BoxConstraints(
+                minHeight: 40.0,
+                minWidth: 90,
+                maxWidth: width,
+                maxHeight: height),
+            alignment: Alignment.center,
+            child: Text(
+              text,
+              textAlign: TextAlign.center,
+              style: textStyle,
             ),
           ),
         ),
       ),
     );
   }
+}
+
+class Background {
+  Background({this.child, this.backgroundColor = Colors.transparent});
+
+  Widget child;
+  Color backgroundColor;
 }