+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Adicionado NkButton, NkFlatButton, NkTextField.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Primeiro upload